Fix a vertex shader generation bug that crashed Super Monkey Ball. Better shader error handling. Random warning fixes.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4170 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/Core.cpp
@@ -660,13 +660,13 @@ void Callback_VideoCopiedToXFB(bool video_update)
 		double wait_frametime = (1000.0 / VideoInterface::TargetRefreshRate);
 
 		if (Timer.GetTimeDifference() >= wait_frametime * frames)
-			no_framelimit=Timer.GetTimeDifference();
+			no_framelimit = (u32)Timer.GetTimeDifference();
 
 		while (Timer.GetTimeDifference() < wait_frametime * videoupd)
 		{
 			// TODO : This is wrong, the sleep shouldn't be there but rather in cputhread
 			// as it's not based on the fps but on the refresh rate...
-			if (no_framelimit==0)
+			if (no_framelimit == 0)
 				Common::SleepCurrentThread(1);
 		}
 	}
